projects/ai-agent/pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = "A boot.dev project"
authors = ["Evan <evan.allender@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-genai = "^1.20.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
